Move ImportConfig and related classes to separate module
