+++ src/file_manager/import_media.py
@@ -1,21 +1,19 @@
 #!/usr/bin/python3
 """Import photos from media to storage."""
-
+from __future__ import annotations
 
 import argparse
 import logging
 import os
 import re
 import sys
-from dataclasses import dataclass
 from datetime import datetime, timedelta
-from enum import Enum
 from pathlib import Path
 
 import exifread  # type: ignore
-from yaml import Loader, load
 
-from file_manager.file_utils import convert_to_bytes, get_lsblk
+from file_manager.file_utils import get_lsblk
+from file_manager.import_config import ImportConfig, MediaConfig, MediaType
 from file_manager.utils import float2timestamp, timeobj2exif_str
 
 _COMPARE_TIME_DIFF = timedelta(2)  # 2 days
@@ -125,61 +123,6 @@ def compare_files(file_name_1: str, dir_path_1: Path, dir_path_2: Path, file_nam
             f"second file {file_path_2} too old {timeobj2exif_str(mtime_2)}")
         return False
     return exif_time_1 == read_file_time(file_path_2)
-
-
-class MediaType(Enum):
-    PHOTOS = 1
-    VIDEOS = 2
-    DASHCAM = 3
-
-
-@dataclass
-class MediaConfig:
-    name: str
-    label: re.Pattern
-    types: list[MediaType]
-
-
-class ImportConfig:
-    def __init__(self, config_file: Path) -> None:
-        self.config = load(config_file.read_text("utf-8"), Loader=Loader)
-        logging.debug("ImportConfig: %s", self.config)
-
-    @property
-    def storage_regex_list(self) -> list[str]:
-        return self.config.get("storage-config", {}).get(
-                "storage-includes", ["^not-a-disk$"])
-
-    @property
-    def import_roots_list(self) -> list[str]:
-        logging.debug("import_roots_list: %s", self.config.get("storage-config", {}).get("import-locations", {}).values())
-        return list({import_location["root"] for import_location in self.config.get("storage-config", {}).get("import-locations", {}).values()})
-
-    @property
-    def free_space_limits(self) -> dict[str, int]:
-        if "free-space-limit" not in self.config.get("storage-config", {}):
-            return {}
-        free_space_limits = {}
-        free_space_config = self.config["storage-config"]["free-space-limit"]
-        if "percentage" in free_space_config:
-            free_space_limits["percentage"] = int(
-                    free_space_config["percentage"])
-        if "absolute" in free_space_config:
-            free_space_limits["absolute"] = convert_to_bytes(
-                    free_space_config["absolute"])
-
-        return free_space_limits
-
-    @property
-    def media_config(self) -> dict[str, MediaConfig]:
-        if "media-config" not in self.config:
-            return {}
-        return {
-            details["label"]: MediaConfig(details["name"], re.compile(details["label"]),
-                                          [MediaType[str(media_type).upper()]
-                                           for media_type in details["types"]])
-            for details in self.config["media-config"]
-        }
 
 
 class MediaFiles:
