+++ test_data/min_import_config.yaml
@@ -1,0 +1,2 @@
+storage-config:
+  storage-includes: []
